Hotfix 0.9.1
Fixes issue with non-ASCII characters in DuraCloud content ID.
The fix uses Addressable::URI.parse from the `addressable` gem
to normalize the non-ASCII characters properly. As a result,
manual escaping of spaces was removed (but escaping percent signs
in content IDs was retained because they are assumed not to
signify character codes).